Use iTunes specific author tag in feed
The old approach of using the `author()` Feedgen method didn’t work, because it requires an email address [1]:

> Get or set author data. An author element is a dictionary containing a name, an email address and a URI. Name is mandatory for ATOM, **email is mandatory for RSS**.

As no email address was supplied, this call was ignored by Feedgen and no `<author>` tag was present in the resulting feed.

However, we can use `podcast.itunes_author()` [2] instead, which doesn’t require an email address and uses the `<itunes:author>` tag instead.

[1]: https://python-feedgen.readthedocs.io/en/latest/api.feed.html#feedgen.feed.FeedGenerator.author

[2]: https://python-feedgen.readthedocs.io/en/latest/ext/api.ext.podcast.html#feedgen.ext.podcast.PodcastExtension.itunes_author